chore(components): move odd components to suite package
