+++ packages/components/src/components/ConfirmOnDevice/ConfirmOnDevice.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import styled, { css } from 'styled-components';
-import { Image } from '../../Image/Image';
-import { variables, animations } from '../../../config';
-import { Icon } from '../../Icon/Icon';
+import { Image } from '../Image/Image';
+import { variables, animations } from '../../config';
+import { Icon } from '../Icon/Icon';
 import { DeviceModelInternal } from '@trezor/connect';
 
 enum AnimationDirection {
