+++ packages/components/src/index.ts
@@ -27,10 +27,9 @@ export * from './components/typography/Heading/Heading';
 export * from './components/typography/Paragraph/Paragraph';
 export * from './components/typography/Link/Link';
 export * from './components/Warning/Warning';
-export * from './components/others/Box';
-export * from './components/others/Card';
-export * from './components/others/SecurityCard';
-export * from './components/prompts/ConfirmOnDevice/ConfirmOnDevice';
+export * from './components/Box/Box';
+export * from './components/Card/Card';
+export * from './components/ConfirmOnDevice/ConfirmOnDevice';
 export * from './components/Modal/Modal';
 export * from './components/Modal/Backdrop';
 export * from './components/logos/CoinLogo/CoinLogo';
@@ -43,8 +42,6 @@ export * from './components/Tooltip/Tooltip';
 export * from './components/Timerange/Timerange';
 export * from './components/Truncate/Truncate';
 export * from './components/form/SelectBar/SelectBar';
-export * from './components/HoverAnimation';
-export * from './components/Fade';
 export * from './components/Image/Image';
 export * from './components/Image/images';
 export * from './components/CollapsibleBox';
