+++ packages/suite/src/components/suite/NavigationBar/components/HoverAnimation.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled, { css } from 'styled-components';
 import { transparentize } from 'polished';
-import { variables } from '../config';
+import { variables } from '@trezor/components/src/config';
 
 const Wrapper = styled.div<{ size?: string; isHoverable?: boolean }>`
     position: relative;
