+++ packages/suite/src/components/suite/NavigationBar/components/NavigationActions/components/ActionItem.tsx
@@ -1,15 +1,8 @@
 import React, { useMemo } from 'react';
 import styled from 'styled-components';
-import {
-    useTheme,
-    Icon,
-    IconProps,
-    variables,
-    HoverAnimation,
-    FluidSpinner,
-    Tooltip,
-} from '@trezor/components';
+import { useTheme, Icon, IconProps, variables, FluidSpinner, Tooltip } from '@trezor/components';
 import { FADE_IN } from '@trezor/components/src/config/animations';
+import { HoverAnimation } from '../../HoverAnimation';
 
 const Wrapper = styled.div<{
     $isOpen: ActionItemProps['isOpen'];
